feat: extend print_message to support vim.notify in interactive mode
Builds on the message routing introduced in the previous commit by
adding support for vim.notify when running interactively and the
function is available. This improves the UX of :MasonUpdateAll in Neovim GUIs.

Streamlines message generation by moving the "[mason-update-all]" prefix
into the print_message function for headless and print() modes.
In vim.notify mode, the context is included in the notification title
instead of the message body.

Also addresses the intent of PR #4.