fix: improve output handling for headless mode (#6)
* fix: improve output handling for headless mode

This commit fixes an issue when mason-update-all is called in Neovim
headles mode (nvim --headless) where each outputed message is appended
to the last with no line break in between. It uses `vim.api.nvim_list_uis()` to
detect when in headless mode so interactive mode is not affected.

* feat: extend print_message to support vim.notify in interactive mode

Builds on the message routing introduced in the previous commit by
adding support for vim.notify when running interactively and the
function is available. This improves the UX of :MasonUpdateAll in Neovim GUIs.

Streamlines message generation by moving the "[mason-update-all]" prefix
into the print_message function for headless and print() modes.
In vim.notify mode, the context is included in the notification title
instead of the message body.

Also addresses the intent of PR #4.

---------

Co-authored-by: John Gorman <967905+gorman-john-m@users.noreply.github.com>